[Fix] Structure of files and folders
