+++ client/package.json
@@ -3,20 +3,22 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@chakra-ui/react": "^1.6.7",
-    "@emotion/react": "^11.4.1",
+    "@chakra-ui/react": "^1.6.10",
+    "@emotion/react": "^11.5.0",
     "@emotion/styled": "^11.3.0",
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.4",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "axios": "^0.23.0",
     "braces": "^3.0.2",
     "framer-motion": "^4.1.17",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-scripts": "^1.1.5",
+    "react-router-dom": "^5.3.0",
+    "react-scripts": "^3.4.4",
     "react-twitch-embed": "^2.1.0",
-    "web-vitals": "^1.1.2"
+    "web-vitals": "^2.1.2",
+    "webpack": "^4.42.0"
   },
   "scripts": {
     "start": "react-scripts start",
