+++ client/public/fonts/readme.txt
@@ -1,0 +1,7 @@
+Twitchy.TV FONT
+
+The Twitchy.TV font is a font that replicates the font used for the logo of the website Twitch.
+
+The Twitch logo and the "Glitch" mascot are registered trademarks of Twitch.
+2016 - By MaxiGamer / maxigamer.deviantart.com
+Made with FontStruct.
