+++ client/src/index.js
@@ -2,11 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
-import MainPage from './views/MainPage'
+import App from './App';
 
 ReactDOM.render(
   <React.StrictMode>
-    <MainPage/>
+    <App/>
   </React.StrictMode>,
   document.getElementById('root')
 );
