--- client/src/styleComponents/theme/Theme.js
@@ -1,12 +1,0 @@
-import { extendTheme } from "@chakra-ui/react";
-
-const theme = extendTheme({
-  colors: {
-    brand: {
-      100: "#f7fafc",
-      900: "#1a202c",
-    },
-  },
-});
-
-export default theme;
