+++ package.json
@@ -11,9 +11,10 @@
   "license": "ISC",
   "description": "Visualizing Twitch game streaming data",
   "dependencies": {
-    "concurrently": "^6.2.1",
+    "concurrently": "^6.3.0",
     "express": "^4.17.1",
-    "mongoose": "^6.0.3",
-    "nodemon": "^2.0.12"
+    "mongoose": "^6.0.11",
+    "nodemon": "^2.0.14",
+    "styled-components": "^4.4.0"
   }
 }
